Standardize use of bypass_event() for event handler bypass.
The decision on whether or not to bypass the modeling of a
security event was implemented in the tsem.c:dispatch_event()
function.  The decision point was after the allocation and setup
of the tsem_event structure for the event had been completed.

This represents a significant committment of processing time and
resources to create and initialize a structure that will never be
used.

Remove the check for the event bypass from the dispatch_event()
function and add a call to the tsem.c:bypass_event() function to
each of the security event handlers.

Testing of this change indicated that there was an excessive
check in the bypass_event() function that was resulting in the
loss of events for the root security modeling space when the
system was running tsem_mode 2.  This was no doubt secondary to
the fact that this check was not relevant after its transfer
from its original locatio nto the bypass_event() function.

This commit drops the following check:

if (tsem_mode == EXPORT_ONLY && !ctx->id && !ctx->external)

Which was preventing the capture of security event descriptions
for export only modeling of the root security modeling namespace.

The drop of this check also enabled the ability to drop the
capture of the security modeling context in a function local
pointer value.  The tagging of this event as an inline function
was contemplated but with the anticipation of loadable security
models the decision was made to not tag this function as
inlinable.