Convert to standardized socket output.
Remaining instances of open coded display of fundamental socket
information were converted to use the fs.c:show_socket_info()
function.

All of the key arguments passed to the function were standardized
to use the same names that were used for the arguments to the
various socket based security handlers.