Convert inode attribute handlers to use tsem_dentry.
The tsem_inode_getattr handler takes as an argument a pointer to a
path structure while the tsem_inode_setattr takes a pointer to a
dentry structure.  Both argument types can be retained in a
tsem_dentry structure.

The tsem_path and tsem_inode_cell structures were replaced in the
'out' structure of the tsem_inode_attr_args structure with a
tsem_dentry structure.

The event.c:get_inode_getattr() and event.c:get_inode_setattr()
functions were modified to fill the tsem_dentry structure in the
'out' portion of the tsem_inode_attr_args structure.

The event.c:fill_dentry_path() function was introduced in order
to allow a tsem_dentry structure to be filled from the context of
path structure.  This is to allow the fill_path() function to be
used to generate the pathname rather than the fill_path_dentry()
function that generates a path from a dentry, which on
mountpoints yields a different pathname.

The event.c:free_cell() function was modified to free the path
member of the tsem_dentry structure for these two handlers.

The clauses for these two handlers in the
map.c:get_cell_mapping() function was modified to use the
add_dentry function in order to add the description of the inode
and path to the security state coefficient.

In a similar fashion the fs.c:show_inode_getattr() and
fs.c:show_inode_setattr() functions were modified to use the
show_dentry() function to display the inode and path, rather than
the use of the show_inode() and show_path() primitives.