Allow key value to be specified for inode display.
The objective is to have the security event descriptions
represent as close as possible the arguments to the security
event handlers.

In some cases an inode is representing the 'dir' argument to a
security handler, and in other cases, such as the output of a
dentry value the proper reference to an inode is as an inode.

In order to support this an additional argument was added to the
fs.c:show_inode() function to allow the desired key name for the
inode to be specified to the caller.