Allow specification of key name for IPC permission.
In order to maintain variable naming consistency with the
arguments to the IPC security handlers there was a need to be
able to specify the key name for the JSON encoded IPC permission
structure.

The security_ipc_permission handler has the kern_ipc_perm pointer
named icpc while the remainder of the IPC permission handlers use
a variable name of perm.

Add an argument to the fs.c:show_ipc_cred() function in order to
specify the name of the JSON key to be used for the output of a
security event description for IPC relevant handlers.