Change argument order for tsem_mmap_file handler.
The security state description for the event was not being
displayed in the order that the arguments were supplied to the
handler.

The fs.c:show_mmap() function was updated to display the events
ordered in the CELL description in the same order as the order
they are passed into the security handler.  With JSON encoding
there is no need to have an explicit field that indicates whether
or not the mapping was anonymous or file backed, since the JSON
parser can detect the present of the file key.  Use the anonymous
flag simply to determine if the file description should be
output.

The map.c:add_mmap_file() was introduced in order to encapsulate
the generation of the security state coefficient.  The parameters
were arranged to be fed into the coefficient generation stream in
the order they are specified in the security handler.  If the
mapping is anonymous the file description is not fed into the
stream.