Update implementation of tsem_task_kill.
The order in which the parameters of the CELL description are
displayed was modified in the fs.c:show_task_kill() function was
modified to match the order that the parameters were specified in
the handler.  The flag indicating whether the signal was a cross
model signal was moved to the end of the list since it is a
parameter added by TSEM.

The map.c:add_task_kill() function was added to encapsulate the
generation of the security state coefficient for the event.  The
open coding of the calls to the hash function were replaced with
helper functions that add the various parameters.