Update tsem_key_permission handler.
One of the arguments to the tsem_key_permission handler is a
key_ref_t that is an anonymous type that resolves into a key
definition structure.  Update the fs.c:show_key_permission()
to display the relevant members of the structure in a JSON
structure with a key name of key_ref.

Add the map.c:add_key_permission() function encapsulate the
creation of the security state coefficient.