Encapsulate dev argument for tsem_inode_mknod.
The standard in the fs.c:show_path() function was to have the
'dev' or device number argument encapsulated in a structure that
breaks out the major/minor numbers.

Convert the fs.c:show_inode_mknod() function to encapsulate the
device number argument to the tsem_inode_mknod handler as a 'dev'
structure.