Implement handler for security_ptrace_access_check.
A review was conducted on security events that were not yet
handled by TSEM.  Certain higher priority handlers will be
implemented before the next release.

The tsem_ptrace_access_check was implemented in order to handle
the other access check that determines whether or not a process
can trace another process.  This is a companion check to the
one implemented by tsem_ptrace_traceme.

The implementation was straight forward.

The string 'ptrace_access_check' was added to the tsem_names
array along with a matching entry for the definition of the
tsem_event_type enumeration.

The tsem.c:tsem_ptrace_access_check() function was added, along
with a hook definition, to serve as the handler for the event.
Like the other task relevant events this handler is called under
the task list so it was implemented as a LOCKED TSEM event.

The map.c:add_ptrace_access_check() function was implemented to
generate the security state coefficient for the event.  This was
mated with a new clause in the switch statement in the
get_cell_mapping() function.

In a similar fashion a new clause was added to the switch
statement in the fs.c:tsme_fs_show_trajectory() function to
call the fs.c:show_ptrace_access_check() function to generate the
security state coefficient for the event.